Move visibility logic to utils
